requirements.txt
openai
pydub
librosa
requests
python-dotenv
mutagen
numpy
scipy
matplotlib
aiohttp
asyncio-mqtt